web_app_azure/requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
adal==1.2.3
applicationinsights==0.11.9
azure-common==1.1.25
azure-core==1.5.0
azure-graphrbac==0.61.1
azure-identity==1.2.0
azure-mgmt-authorization==0.60.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-network==10.2.0
azure-mgmt-resource==9.0.0
azure-mgmt-storage==10.0.0
azureml-automl-core==1.6.0
azureml-core==1.6.0
azureml-dataprep==1.6.3
azureml-dataprep-native==14.1.0
azureml-pipeline==1.6.0
azureml-pipeline-core==1.6.0
azureml-pipeline-steps==1.6.0
azureml-sdk==1.6.0
azureml-telemetry==1.6.0
azureml-train==1.6.0
azureml-train-automl-client==1.6.0.post1
azureml-train-core==1.6.0.post1
azureml-train-restclients-hyperdrive==1.6.0
backports.tempfile==1.0
backports.weakref==1.0.post1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==6.7
cloudpickle==1.4.1
contextlib2==0.6.0.post1
cryptography==2.9.2
distro==1.5.0
docker==4.2.0
dotnetcore2==2.1.14
entrypoints==0.3
flake8==3.7.9
Flask==1.0.2
fusepy==3.0.1
idna==2.9
importlib-metadata==1.6.0
isodate==0.6.0
itsdangerous==0.24
jeepney==0.4.3
Jinja2==2.10
jmespath==0.10.0
jsonpickle==1.4.1
MarkupSafe==1.0
mccabe==0.6.1
msal==1.3.0
msal-extensions==0.1.3
msrest==0.6.14
msrestazure==0.6.3
ndg-httpsclient==0.5.1
oauthlib==3.1.0
pathspec==0.8.0
portalocker==1.7.0
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
SecretStorage==3.1.2
six==1.15.0
urllib3==1.25.9
websocket-client==0.57.0
Werkzeug==0.14.1
zipp==3.1.0

